+++ DatabaseLibrary/Database/Reviews/ReviewDTO.cs
@@ -13,5 +13,5 @@ public class ReviewDTO
     public int ReviewerID { get; set; }
     public DateTime DatePosted { get; set; }
     public string ReviewText { get; set; }
-    public float ReviewStars { get; set; }
+    public int ReviewStars { get; set; }
 }
